add invertible module with Invertible trait
